Regression refractoring / comments ++
